refactor(avalonminer): inline summary helpers per review feedback
